+++ android/src/main/java/com/jhotadhari/reactnative/mapsforge/vtm/PathLayerManager.java
@@ -439,7 +439,7 @@ public class PathLayerManager extends LayerManager<PathLayerManager.PathEntry> {
 				styleBuilder.buffer(styleMap.getDouble("buffer"));
 			}
 			if (styleMap.hasKey("scalingZoomLevel")) {
-				styleBuilder.scaleZoomLevel(styleMap.getInt("scalingZoomLevel"));
+				styleBuilder.scaleZoomLevel((int) styleMap.getDouble("scalingZoomLevel"));
 			}
 			if (styleMap.hasKey("cap")) {
 				Paint.Cap cap = switch (Objects.requireNonNull(styleMap.getString("cap"))) {
