+++ android/src/main/java/com/jhotadhari/reactnative/mapsforge/vtm/ShapeLayerManager.java
@@ -355,7 +355,7 @@ public class ShapeLayerManager extends LayerManager<ShapeLayerManager.ShapeEntry
 				styleBuilder.buffer(styleMap.getDouble("buffer"));
 			}
 			if (styleMap.hasKey("scalingZoomLevel")) {
-				styleBuilder.scaleZoomLevel(styleMap.getInt("scalingZoomLevel"));
+				styleBuilder.scaleZoomLevel((int) styleMap.getDouble("scalingZoomLevel"));
 			}
 			if (styleMap.hasKey("cap")) {
 				Paint.Cap cap = switch (Objects.requireNonNull(styleMap.getString("cap"))) {
