+++ src/NativeModules/NativeLayerBitmapTile.ts
@@ -8,10 +8,10 @@ import type { ErrorBase, ResponseBase } from '../types';
 export interface ModuleParams {
 	url?: string;
 	alpha?: Double; // float between 0 and 1.
-	zoomMin?: Int32;
-	zoomMax?: Int32;
-	enabledZoomMin?: Int32;
-	enabledZoomMax?: Int32;
+	zoomMin?: Double;
+	zoomMax?: Double;
+	enabledZoomMin?: Double;
+	enabledZoomMax?: Double;
 	cacheSize?: Int32; // mb
 	cacheDirBase?: string; // empty will be handled java side.
 	cacheDirChild?: string; // empty will be handled java side.
@@ -30,8 +30,8 @@ interface RemoveLayerParams {
 interface UpdateEnabledZoomMinMaxParams {
 	nativeNodeHandle: Int32;
 	uuid: string;
-	enabledZoomMin?: Int32;
-	enabledZoomMax?: Int32;
+	enabledZoomMin?: Double;
+	enabledZoomMax?: Double;
 }
 
 interface SetAlphaParams {
