+++ src/NativeModules/NativeLayerPathJts.ts
@@ -38,7 +38,7 @@ export type GeometryStyleJts = {
 	fillColor?: `#${string}`;
 	fillAlpha?: Double;
 	buffer?: Double;
-	scalingZoomLevel?: Int32;
+	scalingZoomLevel?: Double;
 	cap?: 'SQUARE' | 'ROUND' | 'BUTT';
 	fixed?: boolean;
 	strokeIncrease?: Double;
@@ -86,7 +86,7 @@ interface UpdateCoordinatesParams {
 		fillColor?: string;
 		fillAlpha?: Double;
 		buffer?: Double;
-		scalingZoomLevel?: Int32;
+		scalingZoomLevel?: Double;
 		cap?: string;
 		fixed?: boolean;
 		strokeIncrease?: Double;
